minor refactoring related to merged peaks; add new test

+++ nsxlib/crystal/UnitCell.cpp
@@ -369,14 +369,14 @@ std::vector<Eigen::Vector3d> UnitCell::generateReflectionsInSphere(double dstarm
     return hkls;
 }
 
-std::vector<Eigen::Vector3d> UnitCell::generateReflectionsInShell(double dmin, double dmax, double wavelength) const
+std::vector<Eigen::Vector3i> UnitCell::generateReflectionsInShell(double dmin, double dmax, double wavelength) const
 {
     const Eigen::Vector3d b1(getReciprocalAVector());
     const Eigen::Vector3d b2(getReciprocalBVector());
     const Eigen::Vector3d b3(getReciprocalCVector());
     Eigen::SelfAdjointEigenSolver<Eigen::Matrix3d> eigen_solver;
     const Eigen::Matrix3d BTB(getReciprocalMetricTensor());
-    std::vector<Eigen::Vector3d> hkls;
+    std::vector<Eigen::Vector3i> hkls;
 
     eigen_solver.compute(BTB);
     double b_min = std::sqrt(eigen_solver.eigenvalues().minCoeff());
@@ -416,7 +416,7 @@ std::vector<Eigen::Vector3d> UnitCell::generateReflectionsInShell(double dmin, d
                     continue;
                 }
 
-                hkls.emplace_back(Eigen::Vector3i(h, k, l).cast<double>());
+                hkls.emplace_back(Eigen::Vector3i(h, k, l));
             }
         }
     }
